functionality for start rating input created
Created a js folder in public for all static js files.

created a js file for the front end logic on the create review web page. This file handles the functionality for the start rating system for the user to rate the books they've read.

if the user clicks on a checkbox/star, the previous stars are also checked and  the stars ahead are unchecked